refactor: se crean dos paquetes para dividir las clases Rest de las Services
