+++ PizzApp/src/main/java/persistencia/servicios/TestService.java
@@ -3,8 +3,9 @@ package persistencia.servicios;
 import modelo.TestUno;
 import org.springframework.transaction.annotation.Transactional;
 import persistencia.repositorios.TestRepository;
+import persistencia.servicios.Service.GenericService;
 
-public class TestService extends GenericService<TestUno>{
+public class TestService extends GenericService<TestUno> {
 
     private TestRepository repository;
