+++ PizzApp/src/main/resources/META-INF/spring-services-context.xml
@@ -14,15 +14,15 @@
         <property name="repository" ref="persistence.testrepository" />
     </bean>
 
-    <bean id="services.pedidoservice" class="persistencia.servicios.PedidoService">
+    <bean id="services.pedidoservice" class="persistencia.servicios.Service.PedidoService">
         <property name="repository" ref="persistence.pedidorepository" />
     </bean>
 
-    <bean id="services.productoservice" class="persistencia.servicios.ProductoService">
+    <bean id="services.productoservice" class="persistencia.servicios.Service.ProductoService">
         <property name="repository" ref="persistence.productorepository" />
     </bean>
 
-    <bean id="services.clienteservice" class="persistencia.servicios.ClienteService">
+    <bean id="services.clienteservice" class="persistencia.servicios.Service.ClienteService">
         <property name="repository" ref="persistence.clienterepository" />
     </bean>
